Fixed Weather error, added header

+++ app/src/main/java/com/example/dinfo/WeatherResponse/Geometry.kt
@@ -6,6 +6,6 @@ import com.google.gson.annotations.SerializedName
 data class Geometry (
 
    @SerializedName("type") var type : String,
-   @SerializedName("coordinates") var coordinates : List<Int>
+   @SerializedName("coordinates") var coordinates : List<Double>
 
 )
